changed title of application to priorities. As well as adding new feature implementations such as hiding floating button when scrolling down and showing it again when scrolling up. Tasks can be deleted with a warning dialog and can be sorted, and moved from pending to done status.

+++ app/src/main/java/com/apps/alemjc/prioritytodo/Storage/TodoStorage.kt
@@ -3,13 +3,13 @@ package com.apps.alemjc.prioritytodo.Storage
 import com.apps.alemjc.prioritytodo.content.Todo
 
 /**
- * Created by alemjc on 11/19/17.
+ * Created by Jean Carlos Henriquez on 11/19/17.
  */
 interface TodoStorage {
 
-    fun create(description:String, priority: Int): Boolean
-    fun update(_id:Long, description: String, priority: Int, status: String):Boolean
-    fun remove(_id: Long):Boolean
-    fun getTodos():ArrayList<Todo>
+    fun create(description: String, priority: Long): Boolean
+    fun update(_id: Long, description: String, priority: Long, status: String): Boolean
+    fun remove(_id: Long): Boolean
+    fun getTodos(): ArrayList<Todo>
 
 }
